server/requirements.txt
fastapi==0.110.3
passlib==1.7.4
pydantic==2.7.1
redis==5.0.1
SQLAlchemy==2.0.29
uvicorn==0.29.0
pymysql==1.1.0
pydantic[email]
python-jose
pillow